requirements.txt
aiohttp==3.7.3
aiohttp-requests==0.1.3
asyncpraw==7.1.0
asyncprawcore==1.5.0
buttons==0.1.9
cryptography==3.3.1
danbot-hosting-py
dblpy==0.4.0
discord-interactive==4.0
discord-py-slash-command==1.0.8.5
discord.py==1.6.0
disputils
dpymenus==1.3.0
emoji==0.6.0
Flask==1.1.2
humanize==3.2.0
jishaku
loremipsum==1.0.5
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
MemeLib
motor==2.3.0
numpy==1.19.4
oauthlib==3.1.0
poetry==1.0.5
pymongo==3.11.2
replit==2.0.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
sr-api
termcolor
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2